% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apparent_prevalence.R
\name{apparent_prevalence}
\alias{apparent_prevalence}
\title{Apparent prevalence from known prevalence}
\usage{
apparent_prevalence(p, sens, spec)
}
\arguments{
\item{p}{the true value of the prevalence}

\item{sens}{the sensitivity of the test}

\item{spec}{the specificity of the test}
}
\value{
the expected value of apparent prevalence
}
\description{
The observed counts of disease is going to be a binomial but
with the apparent prevalence as a probability. This will never be less than
\code{(1-specificity)} of the test (and never more than the sensitivity). When either
of those quantities are uncertain the shape of the distribution of observed counts
is not clear cut.
}
\examples{
apparent_prevalence(0, 0.75, 0.97)
apparent_prevalence(1, 0.75, 0.97)
}
