% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{as_tibble.beta_dist}
\alias{as_tibble.beta_dist}
\title{convert a beta distribution to a tibble}
\usage{
\method{as_tibble}{beta_dist}(x, prefix = NULL, confint = 0.95, ...)
}
\arguments{
\item{x}{the beta distribution}

\item{prefix}{name to output columns prefix.lower, prefix.upper etc}

\item{confint}{confidence intervals}

\item{...}{not used}
}
\description{
convert a beta distribution to a tibble
}
