% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{as_tibble.beta_dist_list}
\alias{as_tibble.beta_dist_list}
\title{convert a list of betas to a tibble}
\usage{
\method{as_tibble}{beta_dist_list}(x, ...)
}
\arguments{
\item{x}{a beta dist list}

\item{...}{
  Arguments passed on to \code{\link[=as_tibble.beta_dist]{as_tibble.beta_dist}}
  \describe{
    \item{\code{prefix}}{name to output columns prefix.lower, prefix.upper etc}
    \item{\code{confint}}{confidence intervals}
  }}
}
\value{
a tibble
}
\description{
convert a list of betas to a tibble
}
