% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-models.R
\name{bayesian_component_simpler_model}
\alias{bayesian_component_simpler_model}
\title{Bayesian simpler model true prevalence for component}
\usage{
bayesian_component_simpler_model(
  pos_obs,
  n_obs,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  ...,
  sens = NULL,
  spec = NULL,
  confint = 0.95,
  fmt = "\%1.2f\%\% [\%1.2f\%\% — \%1.2f\%\%]",
  chains = 4,
  warmup = 1000,
  iter = 2000
)
}
\arguments{
\item{pos_obs}{a vector of the number of positive observations for each component test}

\item{n_obs}{a vector of the number of observations for each component test}

\item{false_pos_controls}{the number of positives that appeared in the
component tests of a specificity disease-free control group. These are by
definition false positives.}

\item{n_controls}{the number of controls in the specificity (optional)
disease-free control group (optional).}

\item{false_neg_diseased}{the number of negatives that appeared in the
component tests of a sensitivity confirmed disease group. These are by
definition false negatives. (optional)}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group (potional)}

\item{...}{not used}

\item{sens}{the prior sensitivity of the components as a \code{beta_dist_list}
(optional)}

\item{spec}{the prior specificity of the components as a \code{beta_dist_list}
(optional)}

\item{confint}{confidence limit width}

\item{fmt}{a \code{sprintf} formatting string accepting 3 numbers}

\item{chains}{A positive integer specifying the number of Markov chains. 
    The default is 4.}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin)
    iterations per chain. If step-size adaptation is on (which it is by default), 
    this also controls the number of iterations for which adaptation is run (and
    hence these warmup samples should not be used for inference). The number of 
    warmup iterations should be smaller than \code{iter} and the default is
    \code{iter/2}.}

\item{iter}{A positive integer specifying the number of iterations for each 
    chain (including warmup). The default is 2000.}
}
\value{
a list of dataframes containing the prevalence, sensitivity, and
specificity estimates, and a \code{stanfit} object with the raw fit data
}
\description{
Bayesian simpler model true prevalence for component
}
