% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-models.R
\name{bayesian_panel_complex_model}
\alias{bayesian_panel_complex_model}
\title{Bayesian models true prevalence for panel}
\usage{
bayesian_panel_complex_model(
  test_results = testerror:::.input_data,
  panel_results = testerror:::.input_panel_data,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  ...,
  sens = NULL,
  spec = NULL,
  panel_sens = NULL,
  panel_spec = NULL,
  panel_name = "Panel",
  confint = 0.95,
  fmt = "\%1.2f\%\% [\%1.2f\%\% — \%1.2f\%\%]",
  chains = 4,
  warmup = 1000,
  iter = 2000
)
}
\arguments{
\item{test_results}{A dataframe containing the following columns:
\itemize{
\item id (character) - the patient identifier
\item result (logical) - the test result
\item test (factor) - the test type
Grouped by: \if{html}{\out{<none>}}
}}

\item{panel_results}{A dataframe containing the following columns:
\itemize{
\item id (character) - the patient identifier
\item result (logical) - the panel result
Grouped by: \if{html}{\out{<none>}}
}}

\item{false_pos_controls}{the number of positives that appeared in the
component tests of a specificity disease-free control group. These are by
definition false positives.}

\item{n_controls}{the number of controls in the specificity (optional)
disease-free control group (optional).}

\item{false_neg_diseased}{the number of negatives that appeared in the
component tests of a sensitivity confirmed disease group. These are by
definition false negatives. (optional)}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group (potional)}

\item{...}{not used}

\item{sens}{the prior sensitivity of the components as a \code{beta_dist_list}
(optional)}

\item{spec}{the prior specificity of the components as a \code{beta_dist_list}
(optional)}

\item{panel_sens}{the prior sensitivity of the panel as a \code{beta_dist}
(optional)}

\item{panel_spec}{the prior specificity of the panel as a \code{beta_dist}
(optional)}

\item{panel_name}{the name of the panel for combined result}

\item{confint}{confidence limit width}

\item{fmt}{a \code{sprintf} formatting string accepting 3 numbers}

\item{chains}{A positive integer specifying the number of Markov chains. 
    The default is 4.}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin)
    iterations per chain. If step-size adaptation is on (which it is by default), 
    this also controls the number of iterations for which adaptation is run (and
    hence these warmup samples should not be used for inference). The number of 
    warmup iterations should be smaller than \code{iter} and the default is
    \code{iter/2}.}

\item{iter}{A positive integer specifying the number of iterations for each 
    chain (including warmup). The default is 2000.}
}
\value{
a list of dataframes containing the prevalence, sensitivity, and
specificity estimates, and a \code{stanfit} object with the raw fit data
}
\description{
Uses resampling to incorporate uncertainty of sensitivity and specificity
into an estimate of true prevalence from a given value of apparent
prevalence.
}
\details{
This is not vectorised
}
