% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_panel_simpler_model.R
\name{bayesian_panel_simpler_model}
\alias{bayesian_panel_simpler_model}
\title{Bayesian simpler model true prevalence for panel}
\usage{
bayesian_panel_simpler_model(
  panel_pos_obs,
  panel_n_obs,
  pos_obs,
  n_obs,
  test_names,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  ...,
  sens = uniform_prior(),
  spec = uniform_prior(),
  panel_sens = uniform_prior(),
  panel_spec = uniform_prior(),
  panel_name = "Panel",
  confint = 0.95,
  fmt = "\%1.2f\%\% [\%1.2f\%\% — \%1.2f\%\%]",
  chains = 4,
  warmup = 1000,
  iter = 2000
)
}
\arguments{
\item{panel_pos_obs}{the number of positive observations for a given panel of tests}

\item{panel_n_obs}{the number of observations for each component test}

\item{pos_obs}{the number of positive observations for a given test}

\item{n_obs}{the number of observations for a given test}

\item{test_names}{a vector of the component test names in desired order}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{false_neg_diseased}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group.}

\item{...}{not used}

\item{sens}{the prior sensitivity of the test as a \code{beta_dist}.}

\item{spec}{the prior specificity of the test as a \code{beta_dist}.}

\item{panel_sens}{the prior sensitivity of the panel as a \code{beta_dist}
(optional)}

\item{panel_spec}{the prior specificity of the panel as a \code{beta_dist}
(optional)}

\item{panel_name}{the name of the panel for combined result}

\item{confint}{confidence interval limits}

\item{fmt}{a \code{sprintf} formatting string accepting 3 numbers}

\item{chains}{A positive integer specifying the number of Markov chains. 
    The default is 4.}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin)
    iterations per chain. If step-size adaptation is on (which it is by default), 
    this also controls the number of iterations for which adaptation is run (and
    hence these warmup samples should not be used for inference). The number of 
    warmup iterations should be smaller than \code{iter} and the default is
    \code{iter/2}.}

\item{iter}{A positive integer specifying the number of iterations for each 
    chain (including warmup). The default is 2000.}
}
\value{
a list of dataframes containing the prevalence, sensitivity, and
specificity estimates, and a \code{stanfit} object with the raw fit data
}
\description{
Bayesian simpler model true prevalence for panel
}
