% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_panel_prevalence.R
\name{bayesian_panel_true_prevalence_model}
\alias{bayesian_panel_true_prevalence_model}
\title{Execute one of a set of bayesian models}
\usage{
bayesian_panel_true_prevalence_model(
  ...,
  model_type = c("logit", "simpler", "complex")
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=bayesian_panel_complex_model]{bayesian_panel_complex_model}}, \code{\link[=bayesian_panel_simpler_model]{bayesian_panel_simpler_model}}, \code{\link[=bayesian_panel_logit_model]{bayesian_panel_logit_model}}
  \describe{
    \item{\code{test_results}}{A dataframe containing the following columns:
\itemize{
\item id (character) - the patient identifier
\item result (logical) - the test result
\item test (factor) - the test type
}

Ungrouped}
    \item{\code{panel_sens}}{the prior sensitivity of the panel as a \code{beta_dist}
(optional)}
    \item{\code{panel_spec}}{the prior specificity of the panel as a \code{beta_dist}
(optional)}
    \item{\code{panel_name}}{the name of the panel for combined result}
    \item{\code{cache_result}}{save the result of the sampling in memory for the current
session}
    \item{\code{false_pos_controls}}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}
    \item{\code{n_controls}}{the number of controls in the specificity
disease-free control group.}
    \item{\code{false_neg_diseased}}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased}
    \item{\code{n_diseased}}{the number of confirmed disease cases in the sensitivity
control group.}
    \item{\code{sens}}{the prior sensitivity of the test as a \code{beta_dist}.}
    \item{\code{spec}}{the prior specificity of the test as a \code{beta_dist}.}
    \item{\code{confint}}{confidence interval limits}
    \item{\code{fmt}}{a \code{sprintf} formatting string accepting 3 numbers}
    \item{\code{chains}}{A positive integer specifying the number of Markov chains. 
    The default is 4.}
    \item{\code{iter}}{A positive integer specifying the number of iterations for each 
    chain (including warmup). The default is 2000.}
    \item{\code{warmup}}{A positive integer specifying the number of warmup (aka burnin)
    iterations per chain. If step-size adaptation is on (which it is by default), 
    this also controls the number of iterations for which adaptation is run (and
    hence these warmup samples should not be used for inference). The number of 
    warmup iterations should be smaller than \code{iter} and the default is
    \code{iter/2}.}
    \item{\code{panel_pos_obs}}{the number of positive observations for a given panel of tests}
    \item{\code{panel_n_obs}}{the number of observations for each component test}
    \item{\code{test_names}}{a vector of the component test names in desired order}
    \item{\code{pos_obs}}{the number of positive observations for a given test}
    \item{\code{n_obs}}{the number of observations for a given test}
  }}

\item{model_type}{The bayesian model used - one of: \code{logit}, \code{simpler}, \code{complex}; default \code{logit}}
}
\value{
A dataframe containing the following columns:
\itemize{
\item prevalence.label (character) - a fomatted label of the true prevalence estimate with CI
\item prevalence.lower (numeric) - the lower estimate
\item prevalence.median (numeric) - the median estimate
\item prevalence.method (character) - the method of estimation
\item prevalence.upper (numeric) - the upper estimate
\item test (character) - the name of the test or panel
}

Ungrouped
}
\description{
Execute one of a set of bayesian models
}
