% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{beta_dist}
\alias{beta_dist}
\title{Generate a beta distribution out of probabilities, or positive and negative counts}
\usage{
beta_dist(p = n - q, q = n - p, n = p + q, ...)
}
\arguments{
\item{p}{the first shape / the probability or count of success}

\item{q}{(optional) the second shape / the probability or count of failure}

\item{n}{(optional) the number of trials.}

\item{...}{}
}
\value{
either a single \code{beta_dist} object or a list of \code{beta_dist}s
}
\description{
Generate a beta distribution out of probabilities, or positive and negative counts
}
\examples{
beta_dist(c(1,2,3),c(3,2,1))
}
