% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{beta_fit}
\alias{beta_fit}
\title{Fit a beta distribution to data using method of moments}
\usage{
beta_fit(samples, na.rm = FALSE)
}
\arguments{
\item{samples}{a set of probabilities}

\item{na.rm}{should we ignore NA values}
}
\value{
a \code{beta_dist} S3 object fitted to the data.
}
\description{
Fit a beta distribution to data using method of moments
}
\examples{
beta_fit(stats::rbeta(10000,40,60))
beta_fit(stats::rbeta(10000,1,99))
}
