% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{beta_params}
\alias{beta_params}
\title{Generate concave beta distribution parameters from mean and confidence intervals}
\usage{
beta_params(median, lower, upper, confint = 0.95, widen = 1, limit = 1, ...)
}
\arguments{
\item{median}{the median of the probability given}

\item{lower}{the lower ci of the probability given}

\item{upper}{the upper ci of the probability given}

\item{confint}{the ci limits}

\item{widen}{widen the spread of the final beta by this factor}

\item{limit}{the lowest possible value for the shape parameters of the resulting
\code{beta_dist} (1 enforces that the distribution is convex)}

\item{...}{not used}
}
\value{
a list with shape1, shape2 values, and d, p, q and r functions
}
\description{
Generate concave beta distribution parameters from mean and confidence intervals
}
\examples{
beta = beta_params(0.25, 0.1, 0.3)
}
