% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logitnorm.R
\name{ci_to_logitnorm}
\alias{ci_to_logitnorm}
\title{Generate mu and sigma parameters for a logitnormal distribution}
\usage{
ci_to_logitnorm(median, lower, upper, ci = 0.95, fix_median = TRUE, ...)
}
\arguments{
\item{median}{the median of the}

\item{lower}{the lower CI}

\item{upper}{the upper CI}

\item{ci}{the confidence limits}

\item{fix_median}{make the median of the logitnorm be the same as the median given.
This can cause issues when very skewed distributions are used}

\item{...}{not used}
}
\value{
a tibble with mu and sigma columns
}
\description{
The resulting logitnorm distribution will have a set median. The confidence intervals
will not match those provided as they are used as a inter-quartile range.
}
