% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{.input_data}
\alias{.input_data}
\title{Dataframe format for component test results}
\format{
An object of class \code{iface} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
.input_data
}
\description{
A dataframe containing the following columns:
\itemize{
\item id (character) - the patient identifier
\item result (logical) - the test result
\item test (factor) - the test type
}

Ungrouped
}
\keyword{data}
