% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{.input_panel_data}
\alias{.input_panel_data}
\title{Dataframe format for panel test results}
\format{
An object of class \code{iface} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 3 columns.
}
\usage{
.input_panel_data
}
\description{
A dataframe containing the following columns:
\itemize{
\item id (character) - the patient identifier
\item result (logical) - the panel result
}

Ungrouped
}
\keyword{data}
