% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{.output_data}
\alias{.output_data}
\title{Dataframe format for true prevalence results}
\format{
An object of class \code{iface} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.
}
\usage{
.output_data
}
\description{
A dataframe containing the following columns:
\itemize{
\item prevalence.label (character) - a fomatted label of the true prevalence estimate with CI
\item prevalence.lower (numeric) - the lower estimate
\item prevalence.median (numeric) - the median estimate
\item prevalence.method (character) - the method of estimation
\item prevalence.upper (numeric) - the upper estimate
\item test (character) - the name of the test or panel
}

Ungrouped
}
\keyword{data}
