% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{format.beta_dist_list}
\alias{format.beta_dist_list}
\title{Format a beta distribution list}
\usage{
\method{format}{beta_dist_list}(x, ...)
}
\arguments{
\item{x}{the beta distribution list}

\item{...}{not used}
}
\value{
nothing
}
\description{
Format a beta distribution list
}
