% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_p_value.R
\name{fp_p_value}
\alias{fp_p_value}
\title{Significance of an uncertain test result}
\usage{
fp_p_value(
  pos_obs,
  n_obs,
  false_pos_controls,
  n_controls,
  format = "\%1.3g",
  lim = 1e-04,
  bonferroni = NULL,
  ...
)
}
\arguments{
\item{pos_obs}{the number of positive observations for a given test}

\item{n_obs}{the number of observations for a given test}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is \code{(1-specificity)*n_controls}}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{format}{a sprintf fmt string for the p-value}

\item{lim}{a lower value to display}

\item{bonferroni}{the number of simultaneous hypotheses that are being tested}

\item{...}{not used}
}
\value{
a vector of p-values for the count
}
\description{
Calculates a p-value for a count of positive test results based
on false positive (specificity) controls. The null
hypothesis is that the prevalence of the disease is zero.
}
\details{
This p_value does not tell you whether this count can be trusted only if the
prevalence of this disease is significantly more than zero after this
observation.
}
\examples{
# calculate p-values for counts derived from 300 samples
# 10 observations is within noise of test
# 20 observations is unlikely on 1200 observations
fp_p_value(c(10,2,4,3,10,20), 1200, c(0,0,2,0,2,0)+2, 800)

# if the same observations are made against a smaller group then we get 
# a positive result for 10
fp_p_value( c(10,2,4,3,10,20), 1000, c(2,2,4,2,4,2), 800)

tibble::tibble(
  x = c(1,2,5,10,20,40,20,20,20,20,20),
  n = 1000,
  fp_controls = c(0,0,0,0,0,0,0,1,2,3,4)+2,
  n_controls = 800
) \%>\% dplyr::mutate(
  p_value = fp_p_value(x, n, fp_controls, n_controls)
) \%>\% dplyr::glimpse()
}
