% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_signif_level.R
\name{fp_signif_level}
\alias{fp_signif_level}
\title{Identify the minimum number of positive test result observations needed
to be confident the disease has a non-zero prevalence.}
\usage{
fp_signif_level(
  n_obs,
  false_pos_controls,
  n_controls,
  bonferroni = NULL,
  ...,
  spec = NULL
)
}
\arguments{
\item{n_obs}{the number of tests performed.}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is \code{(1-specificity)*n_controls}}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{bonferroni}{the number of simultaneous tests considered.}

\item{...}{not used}

\item{spec}{a prior value for specificity as a \code{beta}}
}
\value{
a vector of test positive counts which are the lowest significant value
that could be regarded as not due to chance.
}
\description{
Identify the minimum number of positive test result observations needed
to be confident the disease has a non-zero prevalence.
}
\examples{
# lowest significant count of positives in 1000 tests 
fp_signif_level(1000, false_pos_controls = 0:5, n_controls=800)
fp_signif_level(c(1000,800,600,400), false_pos_controls = 1:4, n_controls=800)
}
