% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{get_beta_shape}
\alias{get_beta_shape}
\title{Get a parameter of the \code{beta_dist}}
\usage{
get_beta_shape(x, type = c("shape1", "shape2", "conc"))
}
\arguments{
\item{x}{a \code{beta_dist} or \code{beta_dist_list} acting as the prior}

\item{type}{the parameter to extract one of \code{shape1} or \code{shape2} or \code{conc}}
}
\value{
a vector of doubles
}
\description{
Get a parameter of the \code{beta_dist}
}
\examples{
get_beta_shape(beta_dist(shape1=1,shape2=1))
get_beta_shape(beta_dist(shape1=2:5,shape2=1:4))
}
