% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logitnorm.R
\name{inv_logit}
\alias{inv_logit}
\title{The inverse logit function}
\usage{
inv_logit(y)
}
\arguments{
\item{y}{a number between -Inf and Inf}
}
\value{
a number between 0 and 1
}
\description{
The inverse logit function
}
