% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{length.beta_dist}
\alias{length.beta_dist}
\title{Detect the length of a beta distribution}
\usage{
\method{length}{beta_dist}(x, ...)
}
\arguments{
\item{x}{the beta distribution}

\item{...}{not used}
}
\value{
always 1
}
\description{
Detect the length of a beta distribution
}
