% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{length.beta_dist_list}
\alias{length.beta_dist_list}
\title{Detect the length of a beta distribution list}
\usage{
\method{length}{beta_dist_list}(x, ...)
}
\arguments{
\item{x}{the beta distribution list}

\item{...}{not used}
}
\value{
the length of the list
}
\description{
Detect the length of a beta distribution list
}
