% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crude-vaccine-effectiveness.R
\name{odds_ratio_ve}
\alias{odds_ratio_ve}
\title{Calculate a vaccine effectiveness estimate based on an odds ratio}
\usage{
odds_ratio_ve(
  vaccinatedCase,
  unvaccinatedCase,
  vaccinatedControl,
  unvaccinatedControl,
  confint = c(0.025, 0.975)
)
}
\arguments{
\item{vaccinatedCase}{count of disease positive vaccine positive}

\item{unvaccinatedCase}{count of disease positive vaccine negative}

\item{vaccinatedControl}{count of disease negative vaccine positive}

\item{unvaccinatedControl}{count of disease negative vaccine positive}

\item{confint}{the confidence intervals}
}
\value{
a dataframe
}
\description{
This assumes that OR ~ RR which is only true if controls >> cases
The OR method can be used in test negative designs where disease positive
relates to vaccine treatable disease and disease negative relates to non
vaccine treatable disease
}
\examples{
tibble::tibble(
  N_vacc = 42240,
  N_unvacc = 42256,
  N_vacc_pn_pos = 49,
  N_unvacc_pn_pos = 90
) |> dplyr::mutate(
  odds_ratio_ve(N_vacc_pn_pos, N_unvacc_pn_pos, N_vacc-N_vacc_pn_pos, N_unvacc-N_unvacc_pn_pos)
)


}
