% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_performance.R
\name{optimal_performance}
\alias{optimal_performance}
\title{Test optimal performance}
\usage{
optimal_performance(p = NULL, sens = NULL, spec = NULL)
}
\arguments{
\item{p}{the prevalence or apparent prevalence}

\item{sens}{the sensitivity of the test}

\item{spec}{the specificity of the test}
}
\value{
the combination of sensitivity and specificity where apparent prevalence equals true prevalence
}
\description{
For a given combination of prevalence, sensitivity and specificity this gives
the critical threshold at which true prevalence equals apparent prevalence
}
\examples{
optimal_performance(p=0.1, sens=0.75)
optimal_performance(p=0.005, spec=0.9975)
}
