% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unadjusted-calculations.R
\name{panel_prevalence}
\alias{panel_prevalence}
\title{Expected test panel prevalence assuming independence}
\usage{
panel_prevalence(p, na.rm = FALSE)
}
\arguments{
\item{na.rm}{remove NA values?}

\item{spec}{a vector of specificity of the component tests}
}
\value{
a single value for the effective specificity of the combination of
the tests
}
\description{
Expected test panel prevalence assuming independence
}
\examples{
panel_prev(p = rep(0.01,24))
}
