% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unadjusted-calculations.R
\name{panel_sens_estimator}
\alias{panel_sens_estimator}
\title{Estimate test panel combination sensitivity}
\usage{
panel_sens_estimator(ap, sens, spec, na.rm = FALSE)
}
\arguments{
\item{ap}{the apparent prevalence or test positivity (one of p or ap must be given)}

\item{sens}{a vector of sensitivities of the component tests}

\item{spec}{a vector of specificity of the component tests}

\item{na.rm}{remove NA values?}
}
\value{
an effective specificity for the combination of the tests
}
\description{
Estimate the sensitivity of a combination of tests, where the tests are
testing for different conditions and positive results are combined
into a panel using a logical OR. Because false negatives from each component
of a panel can be cancelled out by true positives, or false positives from
other components of the test depending on the prevalence of the underlying
conditions, the combined false negative rate is lower the more cases there
arecombine the false positive rate for the panel is higher than the
individual components (and hence the true negative rate a.k.a specificity is
lower).
}
\examples{
#TODO
}
