% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_spec.R
\name{panel_spec}
\alias{panel_spec}
\title{Test panel combination specificity}
\usage{
panel_spec(spec, na.rm = FALSE)
}
\arguments{
\item{spec}{a vector of specificity of the component tests}

\item{na.rm}{remove NA values?}
}
\value{
a single value for the effective specificity of the combination of
the tests
}
\description{
Calculate the specificity of a combination of tests, where the tests are
testing for different conditions and positive results are combined
into a panel using a logical OR. Because false positives from each component
of a panel combine the false positive rate for the panel is higher than the
individual components (and hence the true negative rate a.k.a specificity is
lower).
}
\examples{
panel_spec(spec = rep(0.9975,24))
}
