% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-error-adjustment.R
\name{prevalence_lang_reiczigel}
\alias{prevalence_lang_reiczigel}
\title{True prevalence from apparent prevalence with uncertainty}
\usage{
prevalence_lang_reiczigel(
  pos_obs,
  n_obs,
  false_pos_controls,
  n_controls,
  false_neg_diseased,
  n_diseased,
  ...,
  confint = 0.95,
  prefix = "prevalence",
  fmt = "\%1.2f\%\% [\%1.2f\%\% — \%1.2f\%\%]"
)
}
\arguments{
\item{pos_obs}{the number of positive observations for a given test}

\item{n_obs}{the number of observations for a given test}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{false_neg_diseased}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group.}

\item{...}{not used}

\item{confint}{confidence interval limits}

\item{prefix}{column name prefix}

\item{fmt}{a \code{sprintf} formatting string accepting 3 numbers}
}
\value{
the expected value of apparent prevalence
}
\description{
Uses lang-reiczigel estimators to incorporate uncertainty of sensitivity and specificity into
an estimate of true prevalence from a given value of apparent prevalence.
}
