% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_panel_lang_reiczigel.R
\name{prevalence_panel_lang_reiczigel}
\alias{prevalence_panel_lang_reiczigel}
\title{Lang-Reiczigel true prevalence for panel}
\usage{
prevalence_panel_lang_reiczigel(
  panel_pos_obs,
  panel_n_obs,
  pos_obs,
  n_obs,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  ...,
  spec = spec_prior(),
  sens = sens_prior(),
  samples = 1000
)
}
\arguments{
\item{panel_pos_obs}{the number of positive observations for a given panel of tests}

\item{panel_n_obs}{a vector of the number of observations for each component test}

\item{pos_obs}{the number of positive observations for a given test}

\item{n_obs}{the number of observations for a given test}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{false_neg_diseased}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group.}

\item{...}{
  Arguments passed on to \code{\link[=prevalence_lang_reiczigel]{prevalence_lang_reiczigel}}
  \describe{
    \item{\code{confint}}{confidence interval limits}
    \item{\code{fmt}}{a \code{sprintf} formatting string accepting 3 numbers}
  }}

\item{spec}{the prior specificity of the test as a \code{beta_dist}.}

\item{sens}{the prior sensitivity of the test as a \code{beta_dist}.}

\item{samples}{number of random draws of sensitivity and specificity (optional - default 1000)}
}
\value{
the expected value of apparent prevalence
}
\description{
Uses resampling to incorporate uncertainty of sensitivity and specificity into
an estimate of true prevalence from a given value of apparent prevalence.
}
\details{
This is not vectorised
}
\examples{
#TODO
}
