% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{print.beta_dist_list}
\alias{print.beta_dist_list}
\title{Print a beta distribution}
\usage{
\method{print}{beta_dist_list}(x, ...)
}
\arguments{
\item{x}{the beta distribution}

\item{...}{not used}
}
\value{
nothing
}
\description{
Print a beta distribution
}
