% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crude-vaccine-effectiveness.R
\name{relative_risk_ve}
\alias{relative_risk_ve}
\title{Calculate a vaccine effectiveness estimate based on a risk ratio}
\usage{
relative_risk_ve(
  vaccinatedCase,
  unvaccinatedCase,
  vaccinatedControl,
  unvaccinatedControl,
  confint = c(0.025, 0.975)
)
}
\arguments{
\item{vaccinatedCase}{count of disease positive vaccine positive}

\item{unvaccinatedCase}{count of disease positive vaccine negative}

\item{vaccinatedControl}{count of disease negative vaccine positive}

\item{unvaccinatedControl}{count of disease negative vaccine positive}

\item{confint}{the confidence intervals}
}
\value{
a dataframe
}
\description{
The RR method cannot be used in test negative designs where disease positive
relates to vaccine treatable disease and disease negative relates to non
vaccine treatable disease. It is only relevant in prospective designs with
a vaccinated and unvaccinated group.
}
\examples{
tibble::tibble(
  N_vacc = 42240,
  N_unvacc = 42256,
  N_vacc_pn_pos = 49,
  N_unvacc_pn_pos = 90
) \%>\% dplyr::mutate(
  relative_risk_ve(N_vacc_pn_pos, N_unvacc_pn_pos, N_vacc-N_vacc_pn_pos, N_unvacc-N_unvacc_pn_pos)
)

# dplyr::bind_rows(lapply(
#   c("katz.log", "adj.log", "bailey", "koopman", "noether",  "sinh-1", "boot"),
#   function(m) {tibble::as_tibble(
#     1-DescTools::BinomRatioCI(N_vacc_pn_pos, N_vacc, N_unvacc_pn_pos, N_unvacc, method = m)
#   ) \%>\% dplyr::mutate(
#     method = m
#   )}
# ))
}
