% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{rep.beta_dist}
\alias{rep.beta_dist}
\title{Repeat a \code{beta_dist}}
\usage{
\method{rep}{beta_dist}(x, times)
}
\arguments{
\item{x}{a \code{beta_dist}}

\item{times}{n}
}
\value{
a \code{beta_dist_list}
}
\description{
Repeat a \code{beta_dist}
}
