% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogan-gladen.R
\name{rogan_gladen}
\alias{rogan_gladen}
\title{True prevalence from apparent prevalence}
\usage{
rogan_gladen(ap, sens, spec)
}
\arguments{
\item{ap}{the expected apparent prevalence.}

\item{sens}{the sensitivity of the test}

\item{spec}{the specificity of the test}
}
\value{
the estimate of 'true prevalence'
}
\description{
This estimator runs into problems with small AP as the Rogan-Gladen
conversion is really using expected apparent prevalence. Getting the expected
value of the AP distribution is complex and the expected value given a single
observation is not in general the ratio of positives / count. The expected
apparent prevalence is never less than the specificity but the observed value
often is. To deal with this the R-G estimator truncates at zero.
}
\examples{
rogan_gladen(50/200, 0.75, 0.97)
}
