% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{sens_prior}
\alias{sens_prior}
\title{The default prior for specificity}
\usage{
sens_prior()
}
\value{
a \code{beta_dist}
}
\description{
If undefined this is 0.70 (0.11 - 1.00). This can be set with \verb{options(testerror.sens_prior = beta_dist(p=??, n=??))}
}
