% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{spec_prior}
\alias{spec_prior}
\title{The default prior for specificity}
\usage{
spec_prior()
}
\value{
a \code{beta_dist}
}
\description{
If undefined this is 0.98 (0.71 - 1.00). This can be set with \verb{options(testerror.spec_prior = beta_dist(p=??, n=??))}
}
