% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-test-error-adjustment.R
\name{true_panel_prevalence}
\alias{true_panel_prevalence}
\title{Calculate an estimate of true prevalence for a single panel and components}
\usage{
true_panel_prevalence(
  test_results = testerror:::.input_data,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  ...,
  sens = NULL,
  spec = NULL,
  panel_name = "Panel",
  confint = 0.95,
  method = c("rogan-gladen", "lang-reiczigel", "bayes"),
  na.rm = TRUE
)
}
\arguments{
\item{test_results}{A dataframe containing the following columns:
\itemize{
\item id (character) - the patient identifier
\item result (logical) - the test result
\item test (factor) - the test type
Grouped by: \if{html}{\out{<none>}}
}}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls (optional if \code{spec} is given)}

\item{n_controls}{the number of controls in the specificity
disease-free control group. (optional if \code{spec} is given)}

\item{false_neg_diseased}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased  (optional if \code{sens} is given)}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group.  (optional if \code{sens} is given)}

\item{...}{passed onto methods}

\item{sens}{the sensitivity of the component tests of test as a \code{beta_dist_list}.
(optional if \code{false_pos_controls} and \code{n_controls} is given)}

\item{spec}{the specificity of the test as a \code{beta_dist_list}
(optional if \code{false_neg_diseased} and \code{n_diseased} is given)}

\item{panel_name}{the name of the panel for combined result}

\item{confint}{confidence limit width}

\item{method}{one of:
\itemize{
\item "lang-reiczigel": Frequentist confidence limits
\item "bayes": Bayesian analysis
\item "rogan-gladen": Rogan gladen with uncertainty
}}

\item{na.rm}{exclude patients with missing results}
}
\value{
A dataframe containing the following columns:
\itemize{
\item prevalence.label (character) - a fomatted label of the true prevalence estimate with CI
\item prevalence.lower (numeric) - the lower estimate
\item prevalence.median (numeric) - the median estimate
\item prevalence.method (character) - the method of estimation
\item prevalence.upper (numeric) - the upper estimate
\item test (character) - the name of the test or panel
Grouped by: \if{html}{\out{<none>}}
}
}
\description{
Uses apparent prevalence, and uncertain
estimates of test sensitivity and test specificity for the 3 methods
described in Supplementary 2. This function works for a single panel per
dataframe, multiple panels will need to call this function multiple times
in a \code{group_modify}.
}
\examples{
tmp = testerror:::panel_example()
true_panel_prevalence(
  test_results = tmp$samples \%>\% dplyr::select(id,test,result = observed),
  false_pos_controls = tmp$performance$false_pos_controls,
  n_controls = tmp$performance$n_controls,
  false_neg_diseased = tmp$performance$false_neg_diseased,
  n_diseased = tmp$performance$n_diseased,
  method = "rogan-gladen"
)
}
