% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-error-adjustment.R
\name{true_prevalence}
\alias{true_prevalence}
\title{Calculate an estimate of true prevalence from apparent prevalence, and uncertain
estimates of test sensitivity and test specificity}
\usage{
true_prevalence(
  pos_obs,
  n_obs,
  false_pos_controls,
  n_controls,
  false_neg_diseased,
  n_diseased,
  confint = 0.95,
  method = c("lang-reiczigel", "rogan-gladen", "gelman"),
  prefix = "prevalence"
)
}
\arguments{
\item{pos_obs}{the number of positive observations for a given test}

\item{n_obs}{the number of observations for a given test}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{false_neg_diseased}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group.}

\item{confint}{confidence limit width}

\item{method}{one of:
\itemize{
\item "lang-reiczigel": Frequentist confidence limits
\item "gelman": Bayesian analysis
}}
}
\value{
a dataframe with prevalence.lower, prevalence.median and prevalence.upper columns
}
\description{
Calculate an estimate of true prevalence from apparent prevalence, and uncertain
estimates of test sensitivity and test specificity
}
\examples{
true_prevalence(c(1:50), 200, 2, 800, 25, 75)
true_prevalence(c(1:10)*2, 200, 25, 800, 4.5, 5.5)
}
