% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertain_panel_rogan_gladen.R
\name{uncertain_panel_rogan_gladen}
\alias{uncertain_panel_rogan_gladen}
\title{Rogan-Gladen true prevalence for panel with resampling}
\usage{
uncertain_panel_rogan_gladen(
  panel_pos_obs,
  panel_n_obs,
  pos_obs,
  n_obs,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  ...,
  sens = sens_prior(),
  spec = spec_prior(),
  confint = 0.95,
  samples = 1000
)
}
\arguments{
\item{panel_pos_obs}{the number of positive observations for a given panel of tests}

\item{panel_n_obs}{the number of observations for each component test}

\item{pos_obs}{the number of positive observations for a given test}

\item{n_obs}{the number of observations for a given test}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{false_neg_diseased}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group.}

\item{...}{
  Arguments passed on to \code{\link[=uncertain_rogan_gladen]{uncertain_rogan_gladen}}
  \describe{
    \item{\code{fmt}}{a \code{sprintf} formatting string accepting 3 numbers}
    \item{\code{seed}}{set seed for reproducibility}
  }}

\item{sens}{the prior sensitivity of the test as a \code{beta_dist}.}

\item{spec}{the prior specificity of the test as a \code{beta_dist}.}

\item{confint}{confidence interval limits}

\item{samples}{number fo random draws of sensitivity and specificity}
}
\value{
the expected value of apparent prevalence
}
\description{
Uses resampling to incorporate uncertainty of sensitivity and specificity into
an estimate of true prevalence from a given value of apparent prevalence.
}
\details{
This is not vectorised
}
