% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertain_panel_spec.R
\name{uncertain_panel_spec}
\alias{uncertain_panel_spec}
\title{Propagate component test specificity into panel specificity}
\usage{
uncertain_panel_spec(
  false_pos_controls = NULL,
  n_controls = NULL,
  ...,
  spec = spec_prior(),
  samples = 1000,
  na.rm = FALSE,
  fit_beta = FALSE
)
}
\arguments{
\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{...}{not used}

\item{spec}{the prior specificity of the test as a \code{beta_dist}.}

\item{samples}{number fo random draws of sensitivity and specificity}

\item{na.rm}{remove missing values}

\item{fit_beta}{return the result as a \code{beta_dist} object?}
}
\value{
a vector of possible specificities for the panel or a fitted \code{beta_dist}
}
\description{
Propagate component test specificity into panel specificity
}
\examples{
uncertain_panel_spec(c(2,3,4,2,2), c(800,800,800,800,800), fit_beta=TRUE)
}
