% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertain_rogan_gladen.R
\name{uncertain_rogan_gladen}
\alias{uncertain_rogan_gladen}
\title{True prevalence from apparent prevalence with uncertainty}
\usage{
uncertain_rogan_gladen(
  pos_obs,
  n_obs,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  ...,
  spec = spec_prior(),
  sens = sens_prior(),
  samples = 1000,
  confint = 0.95,
  fmt = "\%1.2f\%\% [\%1.2f\%\% — \%1.2f\%\%]",
  seed = NA
)
}
\arguments{
\item{pos_obs}{the number of positive observations for a given test}

\item{n_obs}{the number of observations for a given test}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{false_neg_diseased}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group.}

\item{...}{not used}

\item{spec}{the prior specificity of the test as a \code{beta_dist}.}

\item{sens}{the prior sensitivity of the test as a \code{beta_dist}.}

\item{samples}{number fo random draws of sensitivity and specificity}

\item{confint}{confidence interval limits}

\item{fmt}{a \code{sprintf} formatting string accepting 3 numbers}

\item{seed}{set seed for reproducibility}
}
\value{
the expected value of apparent prevalence
}
\description{
Uses resampling to incorporate uncertainty of sensitivity and specificity into
an estimate of true prevalence from a given value of apparent prevalence.
}
\examples{
uncertain_rogan_gladen(
  pos_obs = 20, n_obs = 1000, 
  false_pos_controls = 10, n_controls = 800, 
  false_neg_diseased = 20, n_diseased = 100)
  
uncertain_rogan_gladen(
  pos_obs = 5, n_obs = 1000, 
  sens = beta_dist(p=0.75,n=200), 
  spec = beta_dist(p=0.9975, n=800))

uncertain_rogan_gladen(
  pos_obs = c(5,10), n_obs = c(1000,1000), 
  false_pos_controls = c(2,1), n_controls = c(800,800), 
  false_neg_diseased = c(25,20),n_diseased = c(100,100))

}
