% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{update_posterior}
\alias{update_posterior}
\title{Update the posterior of a \code{beta_dist}}
\usage{
update_posterior(x, ..., pos = NULL, neg = NULL, n = NULL)
}
\arguments{
\item{x}{a \code{beta_dist} or \code{beta_dist_list} acting as the prior}

\item{...}{not used}

\item{pos}{positive observation(s)}

\item{neg}{negative observation(s)}

\item{n}{number observations}
}
\value{
a new \code{beta_dist} o \code{beta_dist_list}
}
\description{
Update the posterior of a \code{beta_dist}
}
\examples{
update_posterior(beta_dist(shape1=1,shape2=1), neg=10, n=30)
}
