% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{format.beta_dist}
\alias{format.beta_dist}
\title{Format a beta distribution}
\usage{
\method{format}{beta_dist}(x, ...)
}
\arguments{
\item{x}{the beta distribution}

\item{...}{not used}
}
\value{
nothing
}
\description{
Format a beta distribution
}
