% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{format.beta_dist}
\alias{format.beta_dist}
\title{Format a beta distribution}
\usage{
\method{format}{beta_dist}(
  x,
  glue = "{sprintf('\%1.3f [\%1.3f–\%1.3f] (N=\%1.2f)',median,lower,upper,conc)}",
  ...
)
}
\arguments{
\item{x}{the beta distribution}

\item{glue}{a glue spec taking any of \code{shape1}, \code{shape2}, \code{conc}, \code{mean}, \code{median}, \code{upper}, \code{lower}}

\item{...}{not used}
}
\value{
nothing
}
\description{
Format a beta distribution
}
\examples{
format(beta_dist(shape1=3,shape2=6), "{format(mean*100, digits=3)}\%")
}
