% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{format.beta_dist_list}
\alias{format.beta_dist_list}
\title{Format a beta distribution list}
\usage{
\method{format}{beta_dist_list}(x, ...)
}
\arguments{
\item{x}{the beta distribution list}

\item{...}{
  Arguments passed on to \code{\link[=format.beta_dist]{format.beta_dist}}
  \describe{
    \item{\code{glue}}{a glue spec taking any of \code{shape1}, \code{shape2}, \code{conc}, \code{mean}, \code{median}, \code{upper}, \code{lower}}
  }}
}
\value{
nothing
}
\description{
Format a beta distribution list
}
