% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_prev.R
\name{panel_prevalence}
\alias{panel_prevalence}
\title{Expected test panel prevalence assuming independence}
\usage{
panel_prevalence(p, na.rm = FALSE)
}
\arguments{
\item{p}{a vector of prevalences of the component tests}

\item{na.rm}{remove NA values?}
}
\value{
a single value for the effective specificity of the combination of
the tests
}
\description{
Expected test panel prevalence assuming independence
}
\examples{
panel_prevalence(p = rep(0.01,24))
}
