% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-test-error-adjustment.R
\name{true_panel_prevalence}
\alias{true_panel_prevalence}
\title{Calculate an estimate of true prevalence for a single panel and components}
\usage{
true_panel_prevalence(
  test_results = interfacer::iface(id = character ~ "the patient identifier", test =
    factor ~ "the test type", result = logical ~ "the test result", .groups = FALSE),
  false_pos_controls = NA,
  n_controls = NA,
  false_neg_diseased = NA,
  n_diseased = NA,
  ...,
  sens = beta_dist(n_diseased - false_neg_diseased, false_neg_diseased),
  spec = beta_dist(n_controls - false_pos_controls, false_pos_controls),
  test_names = levels(test_results$test),
  panel_name = "Panel",
  confint = 0.95,
  method = c("rogan-gladen", "lang-reiczigel", "bayes")
)
}
\arguments{
\item{test_results}{A dataframe containing the following columns:
\itemize{
\item id (character) - the patient identifier
\item result (logical) - the test result
\item test (factor) - the test type
Grouped by: \if{html}{\out{<none>}}
}}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{false_neg_diseased}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group.}

\item{...}{passed onto methods}

\item{sens}{the sensitivity of the component tests of test as a \code{beta_dist_list}}

\item{spec}{the specificity of the test as a \code{beta_dist_list}}

\item{test_names}{a vector of the component test names in desired order (optional)}

\item{panel_name}{the name of the panel for combined result}

\item{confint}{confidence limit width}

\item{method}{one of:
\itemize{
\item "lang-reiczigel": Frequentist confidence limits
\item "bayes": Bayesian analysis
\item "rogan-gladen": Rogan gladen with uncertainty
}}
}
\value{
a dataframe with prevalence.lower, prevalence.median and prevalence.upper columns
}
\description{
Uses apparent prevalence, and uncertain
estimates of test sensitivity and test specificity for the 3 methods
described in Supplementary 2. This function works for a single panel per
dataframe, mulitple panels will need to call this function multiple times
in a \code{group_modify}.
}
\examples{
tmp = random_example()
true_panel_prevalence(
  test_results = tmp$samples \%>\% select(id,test,result = observed),
  false_pos_controls = tmp$performance$false_pos_controls,
  n_controls = tmp$performance$n_controls,
  false_neg_diseased = tmp$performance$false_neg_diseased,
  n_diseased = tmp$performance$n_diseased,
  method = "rogan-gladen"
)
}
