% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_prevalence.R
\name{true_prevalence}
\alias{true_prevalence}
\title{Vectorised true prevalence estimates}
\usage{
true_prevalence(
  pos_obs,
  n_obs,
  false_pos_controls = NULL,
  n_controls = NULL,
  false_neg_diseased = NULL,
  n_diseased = NULL,
  confint = 0.95,
  method = c("lang-reiczigel", "rogan-gladen", "bayes"),
  ...,
  spec = NULL,
  sens = NULL
)
}
\arguments{
\item{pos_obs}{the number of positive observations for a given test}

\item{n_obs}{the number of observations for a given test}

\item{false_pos_controls}{the number of positives that appeared in the specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls}

\item{n_controls}{the number of controls in the specificity
disease-free control group.}

\item{false_neg_diseased}{the number of negatives that appeared in the sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group.}

\item{confint}{confidence interval limits}

\item{method}{one of:
\itemize{
\item "lang-reiczigel": Frequentist confidence limits: see \code{prevalence_lang_reiczigel()}
\item "rogan-gladen": Rogan gladen incorporating uncertainty with resampling: see \code{uncertain_rogan_gladen()}
\item "bayes": Bayesian analysis: see \code{bayesian_component_simpler_model()}
}}

\item{...}{
  Arguments passed on to \code{\link[=uncertain_rogan_gladen]{uncertain_rogan_gladen}}
  \describe{
    \item{\code{samples}}{number fo random draws of sensitivity and specificity}
    \item{\code{fmt}}{a \code{sprintf} formatting string accepting 3 numbers}
    \item{\code{seed}}{set seed for reproducibility}
  }}

\item{spec}{the prior specificity of the test as a \code{beta_dist}.}

\item{sens}{the prior sensitivity of the test as a \code{beta_dist}.}
}
\value{
A dataframe containing the following columns:
\itemize{
\item prevalence.label (character) - a fomatted label of the true prevalence estimate with CI
\item prevalence.lower (numeric) - the lower estimate
\item prevalence.median (numeric) - the median estimate
\item prevalence.method (character) - the method of estimation
\item prevalence.upper (numeric) - the upper estimate
\item test (character) - the name of the test or panel
}

Ungrouped
}
\description{
Calculate an estimate of true prevalence from apparent prevalence, and uncertain
estimates of test sensitivity and test specificity, using one of 3 methods.
}
\examples{
true_prevalence(c(1:50), 200, 2, 800, 25, 75)
true_prevalence(c(1:10)*2, 200, 25, 800, 1, 6, method="rogan-gladen")
true_prevalence(c(1:10)*2, 200, 5, 800, 1, 6, method="bayes")
}
