% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-test-error-adjustment.R
\name{uncertain_panel_rogan_gladen}
\alias{uncertain_panel_rogan_gladen}
\title{Rogan-Gladen true prevalence for panel with resampling}
\usage{
uncertain_panel_rogan_gladen(
  panel_pos_obs,
  panel_n_obs,
  pos_obs,
  n_obs,
  false_pos_controls = NA,
  n_controls = NA,
  false_neg_diseased = NA,
  n_diseased = NA,
  ...,
  sens = beta_dist(n_diseased - false_neg_diseased, false_neg_diseased),
  spec = beta_dist(n_controls - false_pos_controls, false_pos_controls),
  samples = 1000
)
}
\arguments{
\item{panel_pos_obs}{the number of positive observations for a given panel of tests}

\item{panel_n_obs}{a vector of the number of observations for each component test}

\item{pos_obs}{a vector of the number of positive observations for each component of a panel test}

\item{n_obs}{a vector of the number of observations for each component test}

\item{false_pos_controls}{the number of positives that appeared in the component tests of a specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls (ignored if \code{spec} is given)}

\item{n_controls}{the number of controls in the specificity
disease-free control group (ignored if \code{spec} is given).}

\item{false_neg_diseased}{the number of negatives that appeared in the component tests of a sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased (ignored if \code{sens} is given)}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group (ignored if \code{sens} is given)}

\item{...}{}

\item{sens}{the sensitivity of the component tests of test as a \code{beta_dist_list} (optional)}

\item{spec}{the specificity of the test as a \code{beta_dist_list} (optional)}

\item{samples}{number of random draws of sensitivity and specificity (optional - default 1000)}
}
\value{
the expected value of apparent prevalence
}
\description{
Uses resampling to incorporate uncertainty of sensitivity and specificity into
an estimate of true prevalence from a given value of apparent prevalence.
}
\details{
This is not vectorised
}
\examples{
#TODO uncertain_panel_rogan_gladen(pos_obs = 20, n_obs = 1000, false_pos_controls = 10, n_controls = 800, false_neg_diseased = 20, n_diseased = 100)
#TODO uncertain_rogan_gladen(pos_obs = 5, n_obs = 1000, sens = beta_dist(0.75,n=200), spec = beta_dist(0.9975, n=800))
}
