% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unadjusted-calculations.R
\name{uncertain_panel_sens_estimator}
\alias{uncertain_panel_sens_estimator}
\title{Propagate component test sensitivity and specificity into panel specificity
assuming a known set of observations of component apparent prevalence}
\usage{
uncertain_panel_sens_estimator(
  pos_obs,
  n_obs,
  false_pos_controls = NA,
  n_controls = NA,
  false_neg_diseased = NA,
  n_diseased = NA,
  ...,
  sens = beta_dist(n_diseased - false_neg_diseased, false_neg_diseased),
  spec = beta_dist(n_controls - false_pos_controls, false_pos_controls),
  samples = 1000,
  confint = 0.95,
  fmt = "\%1.2f\%\% [\%1.2f\%\% — \%1.2f\%\%]",
  seed = NA,
  na.rm = FALSE,
  fit_beta = FALSE
)
}
\arguments{
\item{pos_obs}{a vector of the number of positive observations for each component of a panel test}

\item{n_obs}{a vector of the number of observations for each component test}

\item{false_pos_controls}{the number of positives that appeared in the component tests of a specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls (ignored if \code{spec} is given)}

\item{n_controls}{the number of controls in the specificity
disease-free control group (ignored if \code{spec} is given).}

\item{false_neg_diseased}{the number of negatives that appeared in the component tests of a sensitivity
confirmed disease group. These are by definition false negatives. This
is (1-sensitivity)*n_diseased (ignored if \code{sens} is given)}

\item{n_diseased}{the number of confirmed disease cases in the sensitivity
control group (ignored if \code{sens} is given)}

\item{...}{}

\item{sens}{the sensitivity of the component tests of test as a \code{beta_dist_list} (optional)}

\item{spec}{the specificity of the test as a \code{beta_dist_list} (optional)}

\item{samples}{number of random draws of sensitivity and specificity (optional - default 1000)}

\item{fit_beta}{return the result as a \code{beta_dist} object?}
}
\value{
a vector of possible sensitivity values
}
\description{
Propagate component test sensitivity and specificity into panel specificity
assuming a known set of observations of component apparent prevalence
}
\examples{
uncertain_panel_sens_estimator(
  pos_obs = c(30,10,20,10,5), n_obs=1000,
  false_pos_controls = c(20,15,15,15,15), n_controls = c(800,800,800,800,800),
  false_neg_diseased = c(20,25,20,20,15), n_diseased = c(100,100,100,100,100),
  fit_beta = TRUE)
}
