% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unadjusted-calculations.R
\name{uncertain_panel_spec}
\alias{uncertain_panel_spec}
\title{Propagate component test specificity into panel specificity}
\usage{
uncertain_panel_spec(
  false_pos_controls = NA,
  n_controls = NA,
  ...,
  spec = beta_dist(n_controls - false_pos_controls + 1, false_pos_controls + 1),
  samples = 1000,
  na.rm = FALSE,
  fit_beta = FALSE
)
}
\arguments{
\item{false_pos_controls}{the number of positives that appeared in the component tests of a specificity
disease-free control group. These are by definition false positives. This
is (1-specificity)*n_controls (ignored if \code{spec} is given)}

\item{n_controls}{the number of controls in the specificity
disease-free control group (ignored if \code{spec} is given).}

\item{...}{not used}

\item{spec}{the specificity of the test as a \code{beta_dist_list} (optional)}

\item{samples}{number of random draws of sensitivity and specificity (optional - default 1000)}

\item{na.rm}{remove missing values}

\item{fit_beta}{return the result as a \code{beta_dist} object?}
}
\value{
a vector of possible specificities for the panel or a fitted \code{beta_dist}
}
\description{
Propagate component test specificity into panel specificity
}
\examples{
uncertain_panel_spec(c(2,3,4,2,2), c(800,800,800,800,800), fit_beta=TRUE)
}
