% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underestimation_threshold.R
\name{underestimation_threshold}
\alias{underestimation_threshold}
\title{Test underestimation limit}
\usage{
underestimation_threshold(sens, spec)
}
\arguments{
\item{sens}{the sensitivity of the test}

\item{spec}{the specificity of the test}
}
\value{
the value where apparent prevalence equals true prevalence
}
\description{
For a given sensitivity and specificity this give the critical threshold after which
test error introduces underestimation rather than over estimation
}
\examples{
tmp1 = underestimation_threshold(0.75, 0.97)
tmp2 = rogan_gladen(tmp1, 0.75, 0.97)
if (abs(tmp1-tmp2) > 0.0000001) stop("error")

}
