% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-distribution.R
\name{update_posterior}
\alias{update_posterior}
\title{Update the posterior of a \code{beta_dist}}
\usage{
update_posterior(x, pos, n, ...)
}
\arguments{
\item{x}{a \code{beta_dist} or \code{beta_dist_list} acting as the prior}

\item{pos}{positive observation(s)}

\item{n}{number observations}

\item{...}{not used}
}
\value{
a new \code{beta_dist} o \code{beta_dist_list}
}
\description{
Update the posterior of a \code{beta_dist}
}
\examples{
update_posterior(beta_dist(1,1), 10, 30)
}
